+++ src/tcl/string/length.ts
@@ -1,0 +1,10 @@
+export function length(str: string): number {
+  //  discuss at: https://locutus.io/tcl/length/
+  // original by: Kevin van Zonneveld (https://kvz.io)
+  //   example 1: length('hello')
+  //   returns 1: 5
+  //   example 2: length('')
+  //   returns 2: 0
+
+  return String(str).length
+}